requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Text processing and embeddings
sentence-transformers>=2.2.0
nltk>=3.8.0
spacy>=3.6.0

# Data processing and monitoring
datasets>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0

# Web framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Database and storage
redis>=4.6.0
sqlite3

# Containerization and deployment
docker>=6.1.0

# Alerting and notifications
requests>=2.31.0
smtplib

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
rich>=13.4.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.17.0
